requirements.txt
# Core LangChain packages (latest as of October 2025)
langchain==0.3.27
langchain-anthropic==0.3.21
langchain-community==0.3.30
langchain-core==0.3.78

# Parallel.ai SDK
parallel-web

# Data validation
pydantic>=2.9.2
pydantic-settings>=2.10.1

# Async support
aiohttp==3.10.10

# Environment management
python-dotenv==1.0.1

# CLI and output formatting
rich==13.9.2

# Utilities
requests>=2.32.5

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# Development
black==24.10.0
flake8==7.1.1
mypy==1.11.2
